refactor(svelte): extract theme store
